Allow opening unpublishable files
