+++ src/pages/registration/files_and_license_tab/UnpublishableFileRow.tsx
@@ -1,5 +1,5 @@
 import CancelIcon from '@mui/icons-material/Cancel';
-import { Checkbox, IconButton, TableCell, TableRow, Tooltip, Typography } from '@mui/material';
+import { Box, Checkbox, IconButton, TableCell, TableRow, Tooltip, Typography } from '@mui/material';
 import { Field, FieldProps, useFormikContext } from 'formik';
 import prettyBytes from 'pretty-bytes';
 import { useState } from 'react';
@@ -9,6 +9,7 @@ import { TruncatableTypography } from '../../../components/TruncatableTypography
 import { AssociatedFile, AssociatedFileType } from '../../../types/associatedArtifact.types';
 import { SpecificFileFieldNames } from '../../../types/publicationFieldNames';
 import { dataTestId } from '../../../utils/dataTestIds';
+import { DownloadFileButton } from './DownloadFileButton';
 
 interface UnpublishableFileRowProps {
   file: AssociatedFile;
@@ -26,28 +27,31 @@ export const UnpublishableFileRow = ({ file, removeFile, baseFieldName, disabled
 
   return (
     <TableRow data-testid={dataTestId.registrationWizard.files.fileRow}>
-      <TableCell sx={{ minWidth: '13rem' }}>
+      <TableCell sx={{ minWidth: '13rem', maxWidth: '20rem' }}>
         <TruncatableTypography>{file.name}</TruncatableTypography>
       </TableCell>
 
       <TableCell>
-        <Tooltip title={t('registration.files_and_license.remove_file')}>
-          <IconButton onClick={toggleOpenConfirmDialog} disabled={disabled}>
-            <CancelIcon color="primary" />
-          </IconButton>
-        </Tooltip>
-        <ConfirmDialog
-          open={openConfirmDialog}
-          title={t('registration.files_and_license.remove_file')}
-          onAccept={() => {
-            removeFile();
-            toggleOpenConfirmDialog();
-          }}
-          onCancel={toggleOpenConfirmDialog}>
-          <Typography>
-            {t('registration.files_and_license.remove_file_description', { fileName: file.name })}
-          </Typography>
-        </ConfirmDialog>
+        <Box sx={{ display: 'flex', alignItems: 'center' }}>
+          <DownloadFileButton file={file} />
+          <Tooltip title={t('registration.files_and_license.remove_file')}>
+            <IconButton onClick={toggleOpenConfirmDialog} disabled={disabled}>
+              <CancelIcon color="primary" />
+            </IconButton>
+          </Tooltip>
+          <ConfirmDialog
+            open={openConfirmDialog}
+            title={t('registration.files_and_license.remove_file')}
+            onAccept={() => {
+              removeFile();
+              toggleOpenConfirmDialog();
+            }}
+            onCancel={toggleOpenConfirmDialog}>
+            <Typography>
+              {t('registration.files_and_license.remove_file_description', { fileName: file.name })}
+            </Typography>
+          </ConfirmDialog>
+        </Box>
       </TableCell>
 
       <TableCell sx={{ minWidth: '5.5rem' }}>{prettyBytes(file.size)}</TableCell>
